Moving windows IO to it's platform home and adding charset conversion to fileCopy
